Move embedding_utils to embed package and make collection_has_embeddings db-only
