+++ lightly_studio/src/lightly_studio/embed/embedding_utils.py
@@ -4,14 +4,17 @@ from uuid import UUID
 
 from sqlmodel import Session
 
-from lightly_studio.dataset.embedding_manager import EmbeddingManagerProvider
 from lightly_studio.resolvers import (
+    collection_embedding_model_resolver,
     sample_embedding_resolver,
 )
 
 
 def collection_has_embeddings(session: Session, collection_id: UUID) -> bool:
     """Check if there are any embeddings available for the given collection.
+
+    Resolves the collection's default embedding model from the database and reports whether
+    any embeddings are stored for it. Does not load the model.
 
     Args:
         session: Database session for resolver operations.
@@ -20,13 +23,12 @@ def collection_has_embeddings(session: Session, collection_id: UUID) -> bool:
     Returns:
         True if embeddings exist for the collection, False otherwise.
     """
-    embedding_manager = EmbeddingManagerProvider.get_embedding_manager()
-    model_id = embedding_manager.load_or_get_default_model(
+    model_id = collection_embedding_model_resolver.get_default_by_collection_id(
         session=session,
         collection_id=collection_id,
     )
     if model_id is None:
-        # No default embedding model loaded for this collection.
+        # No default embedding model registered for this collection.
         return False
 
     return (
